Store money as numeric instead of float
Balance, price and transaction amount were DOUBLE PRECISION, so 19.99
times three came out as 59.97000000000001 and a balance check could
fail on equal amounts. They are Numeric(10,2) now, and the values that
feed them are Decimal end to end.

Output schemas stay float, so the JSON the frontend receives is
unchanged: a number, not a quoted string.

+++ services/payment_service.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from core.enum import TransactionType
 from database.unit_of_work import UnitOfWork
 from schemas.transaction.input_dto import TransactionCreateDTO
@@ -9,7 +11,7 @@ logger = get_logger(__name__)
 class PaymentService:
 
     @staticmethod
-    async def handle_payment_success(client_id: int, amount: float):
+    async def handle_payment_success(client_id: int, amount: Decimal):
         async with UnitOfWork() as uow:
             client = await uow.client.get_client(client_id)
             if client:
